safari-600.7-branch/LayoutTests/fast/dom/shadow/shadowdom-for-media-expected.txt

 
PASS shadowRootForVideoWithControls is youngerShadowRootForVideoWithControls
PASS shadowRootForVideoWithControls is not oldestShadowRootForVideoWithControls
PASS shadowRootForVideoWithoutControls is youngerShadowRootForVideoWithoutControls
PASS shadowRootForVideoWithoutControls is not oldestShadowRootForVideoWithoutControls
PASS successfullyParsed is true

TEST COMPLETE
